+++ models/models.js
@@ -1,17 +1,5 @@
 const express = require("express");
 const app = express();
-const { Sequelize, DataTypes } = require("sequelize");
-require("dotenv").config();
 
-    const seq = new Sequelize('database', process.env.DB_USER, process.env.DB_PASSWORD, {
-        host: process.env.DB_HOST,
-        dialect: "mysql"
-    });
+const priceServiceUrl = ' https://jqjfct0r76.execute-api.us-east-1.amazonaws.com/default/PythonPandas';
 
-const Film = seq.define('database', {
-
-},
-{
-    tableName: 'database',
-    timestamps: false
-});
